refactor(ai): migrate list_alumni to tools registry
