+++ src/lib/ai/tools/executor.ts
@@ -158,16 +158,6 @@ const NON_ADMIN_RLS_READ_TOOL_NAMES: ReadonlySet<ToolName> = new Set<ToolName>([
   "find_navigation_targets",
 ]);
 
-const listAlumniSchema = z
-  .object({
-    limit: z.number().int().min(1).max(25).optional(),
-    graduation_year: z.number().int().min(1900).max(2100).optional(),
-    industry: z.string().trim().min(1).optional(),
-    company: z.string().trim().min(1).optional(),
-    city: z.string().trim().min(1).optional(),
-  })
-  .strict();
-
 const listDonationsSchema = z
   .object({
     limit: z.number().int().min(1).max(25).optional(),
@@ -375,7 +365,6 @@ const findNavigationTargetsSchema = z
   .strict();
 
 const ARG_SCHEMAS: Partial<Record<ToolName, z.ZodSchema>> = {
-  list_alumni: listAlumniSchema,
   list_enterprise_alumni: listEnterpriseAlumniSchema,
   list_donations: listDonationsSchema,
   list_parents: listParentsSchema,
@@ -624,58 +613,6 @@ interface EventValidationErrorRecord {
   index: number;
   missing_fields: string[];
   draft: Record<string, unknown>;
-}
-
-async function listAlumni(
-  sb: SB,
-  orgId: string,
-  args: z.infer<typeof listAlumniSchema>,
-  logContext: AiLogContext
-): Promise<ToolExecutionResult> {
-  const limit = Math.min(args.limit ?? 10, 25);
-  return safeToolQuery(logContext, async () => {
-    let query = sb
-      .from("alumni")
-      .select("id, first_name, last_name, graduation_year, current_company, industry, current_city, position_title, job_title, linkedin_url, email")
-      .eq("organization_id", orgId)
-      .is("deleted_at", null)
-      .order("graduation_year", { ascending: false })
-      .limit(limit);
-
-    if (args.graduation_year !== undefined) {
-      query = query.eq("graduation_year", args.graduation_year);
-    }
-    if (args.industry) {
-      query = query.ilike("industry", `%${sanitizeIlikeInput(args.industry)}%`);
-    }
-    if (args.company) {
-      query = query.ilike("current_company", `%${sanitizeIlikeInput(args.company)}%`);
-    }
-    if (args.city) {
-      query = query.ilike("current_city", `%${sanitizeIlikeInput(args.city)}%`);
-    }
-
-    const { data, error } = await query;
-
-    if (!Array.isArray(data) || error) {
-      return { data, error };
-    }
-
-    return {
-      data: data.map((row) => ({
-        id: row.id,
-        name: buildMemberName(row.first_name ?? "", row.last_name ?? ""),
-        graduation_year: row.graduation_year ?? null,
-        current_company: row.current_company ?? null,
-        industry: row.industry ?? null,
-        current_city: row.current_city ?? null,
-        title: row.position_title ?? row.job_title ?? null,
-        linkedin_url: row.linkedin_url ?? null,
-        email: row.email ?? null,
-      })),
-      error: null,
-    };
-  });
 }
 
 async function listDonations(
@@ -2959,13 +2896,6 @@ export async function executeToolCall(
         return dispatchToolModule(toolName, args, { ctx, sb, logContext });
       }
       switch (toolName) {
-        case "list_alumni":
-          return listAlumni(
-            sb,
-            ctx.orgId,
-            args as z.infer<typeof listAlumniSchema>,
-            logContext
-          );
         case "list_enterprise_alumni":
           return safeToolQuery(logContext, () =>
             listEnterpriseAlumni(
